requirements.txt
pandasai
streamlit
streamlit-chat
streamlit-extras
python-dotenv
langchain
langchain-community
chromadb
pandas

# prompt dışına çıkıldığında uyarı veren kodu yaz